Store board in database, Board reads are now byte (#101)

+++ src/main/java/de/lemonpie/beddomischer/network/reader/CardReadCommand.java
@@ -25,7 +25,7 @@ public class CardReadCommand implements Command
 	@Override
 	public void execute(CommandData command)
 	{
-		int readerId = command.getKey();
+		byte readerId = (byte) command.getKey();
 		String cardCode = command.getValue().getAsJsonPrimitive().getAsString();
 		Card card = Card.fromString(cardCode);
